Added authorization for todos

+++ controllers/todo-controller.js
@@ -1,6 +1,6 @@
 const validate = require('../validation/validate')
 const { todoSchema } = require('../validation/schemas');
-const { ValidationError } = require('../middleware/errors');
+const { ValidationError, ForbiddenError } = require('../middleware/errors');
 const {
     addTodo,
     completeTodo,
@@ -11,15 +11,19 @@ const {
     getAllTodos,
     getTodos
 } = require('../services/').todoServices;
-
+const authorizationSchema = require('../auth/index');
 module.exports = (app) => {
 
     app.get('/todos', async (req, res) => {
+        if (!authorizationSchema.can(req.tokenData.user.role).readAny('todo').granted)
+            throw new ForbiddenError('You are not allowed to read todos')
         res.status(200).json(await getAllTodos());
     });
 
     app.get('/todos/:id', async (req, res, next) => {
         try {
+            if (!authorizationSchema.can(req.tokenData.user.role).readAny('todo').granted)
+                throw new ForbiddenError('You are not allowed to read todos')
             let todo = await getTodoById(req.params.id);
             res.status(200).json(todo)
         } catch (err) {
@@ -31,6 +35,8 @@ module.exports = (app) => {
         try {
             if (req.params.isDone !== 'true' || req.params.isDone !== 'false')
                 throw new ValidationError('isDone must be \'true\' or \'false\'')
+            if (!authorizationSchema.can(req.tokenData.user.role).readAny('todo').granted)
+                throw new ForbiddenError('You are not allowed to read todos')
             let filter = { isCompleted: (req.params.isDone === 'true') };
             res.status(200).json(await getTodos(filter))
         } catch (err) {
@@ -40,6 +46,8 @@ module.exports = (app) => {
 
     app.post('/todos', async (req, res, next) => {
         try {
+            if (!authorizationSchema.can(req.tokenData.user.role).createAny('todo').granted)
+                throw new ForbiddenError('You are not allowed to create todos')
             let todo = await validate(todoSchema, req.body)
             res.status(201).json(await addTodo(todo))
         }
@@ -52,6 +60,8 @@ module.exports = (app) => {
         const updates = req.body
         const id = req.params.id
         try {
+            if (!authorizationSchema.can(req.tokenData.user.role).updateAny('todo').granted)
+                throw new ForbiddenError('You are not allowed to update todos')
             res.status(200).json(await updateTodo(id, updates))
         } catch (err) {
             next(err)
@@ -61,6 +71,8 @@ module.exports = (app) => {
     app.patch('/todos/:id/complete', async (req, res, next) => {
         const id = req.params.id
         try {
+            if (!authorizationSchema.can(req.tokenData.user.role).updateAny('todo').granted)
+                throw new ForbiddenError('You are not allowed to update todos')
             res.status(200).json(await completeTodo(id))
         } catch (err) {
             next(err)
@@ -70,6 +82,8 @@ module.exports = (app) => {
     app.patch('/todos/:id/reset', async (req, res, next) => {
         const id = req.params.id
         try {
+            if (!authorizationSchema.can(req.tokenData.user.role).updateAny('todo').granted)
+                throw new ForbiddenError('You are not allowed to update todos')
             res.status(200).json(await resetTodo(id))
         } catch (err) {
             next(err)
@@ -78,6 +92,8 @@ module.exports = (app) => {
 
     app.delete('/todos/:id', async (req, res, next) => {
         try {
+            if (!authorizationSchema.can(req.tokenData.user.role).deleteAny('todo').granted)
+                throw new ForbiddenError('You are not allowed to delete todos')
             res.status(200).json(await deleteTodo(req.params.id))
         }
         catch (err) {
